requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-aiplatform
ultralytics
onnx
onnxruntime
onnxruntime-tools
